Update the app so it could be operated from the menu bar icon even when the main window is out of focus.
