+++ Helium/Helium/AppDelegate.swift
@@ -67,5 +67,154 @@ class AppDelegate: NSObject, NSApplicationDelegate, NSMenuDelegate {
             
         }
     }
+    
+    var alpha: CGFloat = 0.6 { //default
+        didSet {
+            if translucent {
+                panel.alphaValue = alpha
+            }
+        }
+    }
+    
+    var translucent: Bool = false {
+        didSet {
+            if !NSApplication.sharedApplication().active {
+                panel.ignoresMouseEvents = translucent
+            }
+            if translucent {
+                panel.opaque = false
+                panel.alphaValue = alpha
+            }
+            else {
+                panel.opaque = true
+                panel.alphaValue = 1.0
+            }
+        }
+    }
+    
+    
+    var panel: NSPanel! {
+        get {
+            return (self.defaultWindow as! NSPanel)
+        }
+    }
+    
+    var webViewController: WebViewController {
+        get {
+            return self.defaultWindow?.contentViewController as! WebViewController
+        }
+    }
+    
+    func windowDidLoad() {
+        panel.floatingPanel = true
+        
+        NSNotificationCenter.defaultCenter().addObserver(self, selector: "didBecomeActive", name: NSApplicationDidBecomeActiveNotification, object: nil)
+        NSNotificationCenter.defaultCenter().addObserver(self, selector: "willResignActive", name: NSApplicationWillResignActiveNotification, object: nil)
+        NSNotificationCenter.defaultCenter().addObserver(self, selector: "didUpdateTitle:", name: "HeliumUpdateTitle", object: nil)
+    }
+    
+    //MARK: IBActions
+    
+    @IBAction func translucencyPress(sender: NSMenuItem) {
+        if sender.state == NSOnState {
+            sender.state = NSOffState
+            didDisableTranslucency()
+        }
+        else {
+            sender.state = NSOnState
+            didEnableTranslucency()
+        }
+    }
+    
+    @IBAction func percentagePress(sender: NSMenuItem) {
+        for button in sender.menu!.itemArray{
+            (button as! NSMenuItem).state = NSOffState
+        }
+        sender.state = NSOnState
+        let value = sender.title.substringToIndex(advance(sender.title.endIndex, -1))
+        if let alpha = value.toInt() {
+            didUpdateAlpha(NSNumber(integer: alpha))
+        }
+    }
+    
+    @IBAction func openLocationPress(sender: AnyObject) {
+        didRequestLocation()
+    }
+    
+    @IBAction func openFilePress(sender: AnyObject) {
+        didRequestFile()
+    }
+    
+    //MARK: Actual functionality
+    
+    func didUpdateTitle(notification: NSNotification) {
+        if let title = notification.object as? String {
+            panel.title = title
+        }
+    }
+    
+    func didRequestFile() {
+        
+        let open = NSOpenPanel()
+        open.allowsMultipleSelection = false
+        open.canChooseFiles = true
+        open.canChooseDirectories = false
+        
+        if open.runModal() == NSModalResponseOK {
+            if let url = open.URL {
+                webViewController.loadURL(url)
+            }
+        }
+    }
+    
+    
+    func didRequestLocation() {
+        let alert = NSAlert()
+        alert.alertStyle = NSAlertStyle.InformationalAlertStyle
+        alert.messageText = "Enter Destination URL"
+        
+        let urlField = NSTextField()
+        urlField.frame = NSRect(x: 0, y: 0, width: 300, height: 20)
+        
+        alert.accessoryView = urlField
+        alert.addButtonWithTitle("Load")
+        alert.addButtonWithTitle("Cancel")
+        alert.beginSheetModalForWindow(defaultWindow!, completionHandler: { response in
+            if response == NSAlertFirstButtonReturn {
+                // Load
+                var text = (alert.accessoryView as! NSTextField).stringValue
+                
+                if !(text.lowercaseString.hasPrefix("http://") || text.lowercaseString.hasPrefix("https://")) {
+                    text = "http://" + text
+                }
+                
+                if let url = NSURL(string: text) {
+                    self.webViewController.loadURL(url)
+                }
+            }
+        })
+    }
+    
+    func didBecomeActive() {
+        panel.ignoresMouseEvents = false
+    }
+    
+    func willResignActive() {
+        if translucent {
+            panel.ignoresMouseEvents = true
+        }
+    }
+    
+    func didEnableTranslucency() {
+        translucent = true
+    }
+    
+    func didDisableTranslucency() {
+        translucent = false
+    }
+    
+    func didUpdateAlpha(newAlpha: NSNumber) {
+        alpha = CGFloat(newAlpha.doubleValue) / CGFloat(100.0)
+    }
 }
